refactor: split server.py into handlers/, browser_pool.py (#38)
